Refine terminal env naming and wait helper
